Code cleanup suggested by Eclipse.

+++ org.jenetics/src/main/java/org/jenetics/stat/Mean.java
@@ -41,7 +41,7 @@ import org.jenetics.util.MappedAccumulator;
  *
  * @author <a href="mailto:franz.wilhelmstoetter@gmx.at">Franz Wilhelmstötter</a>
  * @since 1.0
- * @version 1.0 &mdash; <em>$Date: 2012-11-06 $</em>
+ * @version 1.0 &mdash; <em>$Date: 2012-11-15 $</em>
  */
 public class Mean<N extends Number> extends MappedAccumulator<N> {
 
@@ -87,7 +87,7 @@ public class Mean<N extends Number> extends MappedAccumulator<N> {
 			_mean = 0;
 		}
 
-		_mean += (value.doubleValue() - _mean)/(double)(++_samples);
+		_mean += (value.doubleValue() - _mean)/(++_samples);
 	}
 
 	@Override
